backend/requirements.txt
Django==4.2.16
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
psycopg2-binary==2.9.7
cloudinary==1.36.0
django-cloudinary-storage==0.3.0
phonenumber-field==7.1.0
django-phonenumber-field==7.1.0
django-countries==7.5.1
django-allauth==0.57.0
crispy-forms==2.1
crispy-tailwind==0.5.0
whitenoise==6.6.0
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
Pillow==10.1.0
requests==2.31.0